src/main/java/me/choicore/study/springframework/core/discount/impl/RateDiscountPolicy.java
package me.choicore.study.springframework.core.discount.impl;

import me.choicore.study.springframework.core.discount.DiscountPolicy;
import me.choicore.study.springframework.core.member.entity.Grade;
import me.choicore.study.springframework.core.member.entity.Member;

public class RateDiscountPolicy implements DiscountPolicy {

    private int discountPercent = 10;

    @Override
    public int discount(final Member member, final int price) {
        if (member.getGrade() == Grade.VIP) {
            return price * discountPercent / 100;
        } else {
            return 0;
        }
    }
}
